requirements.txt
Package                 Version
----------------------- ------------
absl-py                 1.0.0
ansiwrap                0.8.4
argon2-cffi             21.3.0
argon2-cffi-bindings    21.2.0
attrs                   21.4.0
autoflake               1.4
autopep8                1.6.0
backcall                0.2.0
beautifulsoup4          4.10.0
bleach                  4.1.0
cachetools              5.0.0
certifi                 2021.10.8
cffi                    1.15.0
charset-normalizer      2.0.12
click                   8.0.4
colorcet                3.0.0
cycler                  0.11.0
debugpy                 1.5.1
decorator               5.1.1
defusedxml              0.7.1
easydict                1.9
einops                  0.3.0
entrypoints             0.4
flake8                  4.0.1
fonttools               4.30.0
google-auth             2.6.0
google-auth-oauthlib    0.4.6
grpcio                  1.44.0
idna                    3.3
importlib-metadata      4.2.0
importlib-resources     5.4.0
ipydatawidgets          4.2.0
ipykernel               6.9.1
ipympl                  0.8.8
ipython                 7.32.0
ipython-genutils        0.2.0
ipywidgets              7.6.5
itk-core                5.2.1.post1
itk-filtering           5.2.1.post1
itk-meshtopolydata      0.7.1
itk-numerics            5.2.1.post1
itkwidgets              0.32.1
jedi                    0.18.1
Jinja2                  3.0.3
joblib                  1.1.0
jsonschema              4.4.0
jupyter                 1.0.0
jupyter-client          7.1.2
jupyter-console         6.4.3
jupyter-core            4.9.2
jupyterlab-pygments     0.1.2
jupyterlab-widgets      1.0.2
jupytext                1.13.7
kiwisolver              1.3.2
Markdown                3.3.5
markdown-it-py          1.1.0
MarkupSafe              2.1.0
matplotlib              3.5.1
matplotlib-inline       0.1.3
mccabe                  0.6.1
mdit-py-plugins         0.3.0
mistune                 0.8.4
monai                   0.8.1
nbclient                0.5.13
nbconvert               6.4.4
nbformat                5.2.0
nest-asyncio            1.5.4
nibabel                 3.1.1
notebook                6.4.8
numpy                   1.21.5
oauthlib                3.2.0
opencv-python           4.5.5.64
packaging               21.3
pandas                  1.3.5
pandocfilters           1.5.0
papermill               2.3.4
param                   1.12.0
parso                   0.8.3
pexpect                 4.8.0
pickleshare             0.7.5
Pillow                  9.0.1
pip                     22.0.4
prometheus-client       0.13.1
prompt-toolkit          3.0.28
protobuf                3.19.4
ptyprocess              0.7.0
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pycodestyle             2.8.0
pycparser               2.21
pyct                    0.4.8
pydicom                 2.2.2
pydot                   1.4.1
pyflakes                2.4.0
Pygments                2.11.2
pyparsing               3.0.7
pyrsistent              0.18.1
python-dateutil         2.8.2
pytz                    2021.3
PyYAML                  6.0
pyzmq                   22.3.0
qtconsole               5.2.2
QtPy                    2.0.1
requests                2.27.1
requests-oauthlib       1.3.1
rsa                     4.8
scikit-learn            1.0.2
scipy                   1.7.3
Send2Trash              1.8.0
setuptools              59.5.0
six                     1.16.0
sklearn                 0.0
soupsieve               2.3.1
tenacity                8.0.1
tensorboard             2.8.0
tensorboard-data-server 0.6.1
tensorboard-plugin-wit  1.8.1
tensorboardX            2.1
terminado               0.13.3
testpath                0.6.0
textwrap3               0.9.2
threadpoolctl           3.1.0
toml                    0.10.2
torch                   1.9.0+cu111
torchaudio              0.9.0
torchvision             0.10.0+cu111
tornado                 6.1
tqdm                    4.59.0
traitlets               5.1.1
traittypes              0.2.1
typing_extensions       4.0.1
urllib3                 1.26.8
wcwidth                 0.2.5
webencodings            0.5.1
Werkzeug                2.0.3
wheel                   0.37.1
widgetsnbextension      3.5.2
xpinyin                 0.7.6
zipp                    3.7.0
zstandard               0.17.0
